Enhance containerization for improved database management
Revamped containerization process to integrate pgadmin, facilitating easy web-based access to database information. This update eliminates the need for manual entry of database connection details in pgadmin. Additionally, the Dockerfile now employs the latest version of Java 21 and includes a build step for the project prior to execution. The pom.xml file was updated to accommodate the latest Spring Boot version and other dependencies. The initialization of ModelMapper was removed in favor of using MapStruct interfaces for improved mapping efficiency.